Centralize PostHog signup capture in Analytics
