+++ lib/mtg_friends/application.ex
@@ -17,9 +17,7 @@ defmodule MtgFriends.Application do
       # Start Finch
       {Finch, name: MtgFriends.Finch},
       # Start the Endpoint (http/https)
-      MtgFriendsWeb.Endpoint,
-      # Start the RateLimiter
-      MtgFriendsWeb.RateLimit
+      MtgFriendsWeb.Endpoint
       # Start a worker by calling: MtgFriends.Worker.start_link(arg)
       # {MtgFriends.Worker, arg}
     ]
