--- lib/mtg_friends_web/rate_limit.ex
@@ -1,3 +1,0 @@
-defmodule MtgFriendsWeb.RateLimit do
-  use Hammer, backend: :ets
-end
